Major Update
Animation is fully supported now, All you need is to parse the sprites using the nge sprite parser and put them in a folder named "Sprites" that is in working directory of the engine. The engine will automatically load them to memory upon launching it. Now you are ready to render sprites/animations directly from memory